requirements.txt
argon2-cffi==20.1.0
asgiref==3.2.10
beautifulsoup4==4.9.3
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
commonmark==0.9.1
commonmarkextensions==0.0.5
cssutils==1.0.2
Django==3.1
django-cors-headers==3.5.0
django-dotenv==1.4.2
idna==2.10
mysqlclient==2.0.1
pycparser==2.20
PyJWT==1.7.1
pynliner==0.8.0
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.10
uWSGI==2.0.19.1
